requirements.txt
flask
nose
rednose
pymongo
pyyaml
